Pikabug/requirements.txt
discord.py
python-dotenv>=1.0.0
anthropic>=0.58.2